+++ functions/src/utils/getTickersQuotes.ts
@@ -16,5 +16,7 @@ export const getTickersQuotes = async (tickers: string[]) => {
     httpsAgent
   );
 
+  console.log(' Tickwers ', tickerQuotesResponse);
+
   return tickerQuotesResponse.quoteResponse.result;
 };
